add authorization

+++ server/routes/product.js
@@ -1,10 +1,12 @@
 var express = require('express');
 var router = express.Router();
 var db = require('../model/dbConfig');
-var d = new Date();
-var date = d.toLocaleDateString().split('/');
-var newDate = [date[1], date[0], date[2]];
+// var d = new Date();
+// var date = d.toLocaleDateString().split('/');
+// var newDate = [date[1], date[0], date[2]];
+var middlewareLogin = require('../middlewares/checkLogin');
 
+// router.use(middlewareLogin.checkLogin)
 
 router.get('/', (req, res, next) => {
     var page = parseInt(req.query.page);
@@ -22,7 +24,7 @@ router.get('/', (req, res, next) => {
 })
 
 router.get('/:id', (req, res, next) => {
-    
+
     let id = req.params.id;
     console.log(id);
     db.findById(id)
@@ -45,7 +47,7 @@ router.get('/get/count', (req, res, next) => {
         })
 })
 
-router.post('/', (req, res, next) => {
+router.post('/', middlewareLogin.checkLogin, (req, res, next) => {
     let title = req.body.title;
     db.create({
         title: title,
@@ -60,46 +62,74 @@ router.post('/', (req, res, next) => {
                     data: result
                 })
             })
-            
+
         }).catch(err => {
             console.log(err);
         })
 })
 
-router.put('/:id', (req, res, next) => {
-    let id = req.params.id;
-    let title = req.body.title;
-    db.findByIdAndUpdate(
-        id,
-        {
-            title: title
-        }
-    )
-        .then(() => {
-            db.findById(id).then(result => {
-                res.json({
-                    message: 'Update success',
-                    data: result
+router.put('/:id', middlewareLogin.checkLogin, (req, res, next) => {
+    console.log('res.locals: ', res.locals);
+    if (res.locals === 1) {
+        let id = req.params.id;
+        let title = req.body.title;
+        db.findByIdAndUpdate(
+            id,
+            {
+                title: title
+            }
+        )
+            .then(() => {
+                db.findById(id).then(result => {
+                    res.json({
+                        message: 'Update success',
+                        data: result
+                    })
                 })
-            })
 
+            }).catch(err => {
+                console.log(err);
+            })
+    } else if (res.locals === 3) {
+        res.json({
+            status: 'can not edit',
+            message: 'Nomal user can not edit this record'
+        })
+    } else {
+        res.json({
+            status: 'can not edit',
+            message: 'Guest can not edit this record'
+        })
+    }
+
+})
+
+router.delete('/:id', middlewareLogin.checkLogin, (req, res, next) => {
+    console.log(res.locals);
+    if (res.locals === 1) {
+        let id = req.params.id;
+        console.log(id);
+        db.deleteOne({
+            _id: id
+        }).then(data => {
+            res.json({
+                message: 'Delete success'
+            })
         }).catch(err => {
             console.log(err);
         })
-})
-
-router.delete('/:id', (req, res, next) => {
-    let id = req.params.id;
-    console.log(id);
-    db.deleteOne({
-        _id: id
-    }).then(data => {
+    } else if (res.locals === 3) {
         res.json({
-            message: 'Delete success'
+            status: 'can not edit',
+            message: 'Nomal user can not edit this record'
         })
-    }).catch(err => {
-        console.log(err);
-    })
+    } else if (res.locals === 0){
+        res.json({
+            status: 'can not edit',
+            message: 'Guest can not edit this record'
+        })
+    }
+
 })
